fix first checkstyle issues

+++ analysis/src/main/java/edu/schwabe/sequencesort/analysis/report/OperationMetric.java
@@ -2,40 +2,40 @@ package edu.schwabe.sequencesort.analysis.report;
 
 public final class OperationMetric {
 
-    private final long value;
+  private final long value;
 
-    private final long n;
+  private final long n;
 
-    public OperationMetric(final long value, final long n) {
-        this.value = value;
-        this.n = n;
-    }
+  public OperationMetric(final long value, final long n) {
+    this.value = value;
+    this.n = n;
+  }
 
-    private double doubleValue() {
-        return this.value;
-    }
+  private double doubleValue() {
+    return this.value;
+  }
 
-    private double doubleN() {
-        return this.n;
-    }
+  private double doubleN() {
+    return this.n;
+  }
 
-    public Long value() {
-        return Long.valueOf(this.value);
-    }
+  public Long value() {
+    return this.value;
+  }
 
-    public Double divideByNLnN() {
-        final var doubleN = this.doubleN();
-        return Double.valueOf(this.doubleValue() / (doubleN * Math.log(doubleN)));
-    }
+  public Double divideByNLnN() {
+    final var doubleN = this.doubleN();
+    return this.doubleValue() / (doubleN * Math.log(doubleN));
+  }
 
-    public Double divideByNSqrtN() {
-        final var doubleN = this.doubleN();
-        return Double.valueOf(this.doubleValue() / (doubleN * Math.sqrt(doubleN)));
-    }
+  public Double divideByNSqrtN() {
+    final var doubleN = this.doubleN();
+    return this.doubleValue() / (doubleN * Math.sqrt(doubleN));
+  }
 
-    public Double divideBySqrN() {
-        final var doubleN = this.doubleN();
-        return Double.valueOf(this.doubleValue() / (doubleN * doubleN));
-    }
+  public Double divideBySqrN() {
+    final var doubleN = this.doubleN();
+    return this.doubleValue() / (doubleN * doubleN);
+  }
 
 }
